requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
babel==2.17.0
billiard==4.2.1
binance==0.3
celery==5.4.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.12
crispy-bootstrap5==2024.10
cryptography==44.0.2
dateparser==1.2.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.6
django-allauth==65.4.1
django-bootstrap5==24.3
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.12.0
django-import-export==4.3.6
django-money==3.5.3
djangorestframework==3.15.2
fontawesome==5.10.1.post1
frozenlist==1.5.0
httpie==3.2.4
idna==3.10
kombu==5.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
money==1.3.0
mpesa-api==0.2.3
multidict==6.1.0
ngrok==1.4.0
numpy==2.2.3
olefile==0.47
opencv-python==4.11.0.86
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2-binary==2.9.10
py-moneyed==3.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
PyJWT==2.10.1
pyngrok==7.2.3
PySocks==1.7.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0
yarl==1.18.3
